+++ internal/server/server.go
@@ -1,4 +1,4 @@
-package pkg
+package server
 
 import (
 	"context"
